pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-study-plan-generator"
version = "0.1.0"
description = "Gamified study plans for mastering Agentic DevOps with Microsoft Foundry & GitHub Actions."
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2,<3",
    "httpx>=0.27,<1",
    "streamlit>=1.40,<2",
    "requests>=2.32,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
    "PyGithub>=2,<3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
